% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHymod.R
\name{simFunc_hymod}
\alias{simFunc_hymod}
\title{hymod simulation}
\usage{
simFunc_hymod(
  x,
  rain,
  pet,
  nWarmUp,
  stateVal = c(100, 30, 27, 25, 30, 0, 0, 0)
)
}
\arguments{
\item{x}{parameter values}

\item{rain}{precipitation input (mm/day)}

\item{pet}{potential evapotranspiration (mm/day)}

\item{nWarmUp}{length of warmup period}

\item{stateVal}{(optional) initial states}
}
\value{
Vector of simulated runoff
}
\description{
Simulation of hymod rainfall-runoff model
}
