% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BassRiverData.R
\docType{data}
\name{BassRiverData}
\alias{BassRiverData}
\title{Hydrological data for Bass River catchment in Victoria, Australia}
\format{
List containing numerical vectors for
        precipitation (\code{Rain.mm}),
        potential evapotranspiration (\code{ET.mm}),
        and runoff (\code{Runoff.mm.day}), and
        date vector (\code{Date})
}
\usage{
data(BassRiver)
}
\description{
Streamflow, rainfall and PET data for Bass River catchment
(227219) in Victoria, Australia. Originally obtained from Francis Chiew.
}
\references{
\url{https://github.com/eachonly/Robust-Gauss-Newton-Algorithm},
\url{http://www.bom.gov.au/water/hrs/}
}
\keyword{datasets}
